+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 ### Added
 - rack-attack bans are logged to `log/rack-attack.log` (@mkasztelnik)
 - Show user email in admin users index view (@mkasztelnik)
+- Pipeline step script generation uses ERB templates (@mkasztelnik)
 
 ### Changed
 - Upgrade to ruby 2.4.1 (@mkasztelnik)